+++ Sources/AnyDoor/Services/Providers/BuiltinProviderRegistry.swift
@@ -66,7 +66,6 @@ enum BuiltinProviderRegistry {
             ScreenshotTranslateProvider(),
             TranslateSelectionProvider(),
             ImageConversionProvider(),
-            NewQuicklinkProvider(),
         ]
     }
 }